updated sessions controller and helpers, and rpsec for user features
