+++ Gemfile
@@ -29,8 +29,11 @@ gem 'spring',        group: :development
 # Use ActiveModel has_secure_password
 gem 'bcrypt', '~> 3.1.7'
 
-group :development, :test do 
-	gem 'rspec-rails', '~> 3.0'	
+
+group :development, :test do
+	gem 'rspec-rails', '~> 3.0'
+  gem 'capybara'
+  gem 'launchy'
 end
 
 # Use unicorn as the app server
