+++ Rakefile
@@ -4,3 +4,11 @@
 require File.expand_path('../config/application', __FILE__)
 
 Rails.application.load_tasks
+
+# namespace :db do
+#   namespace :test do
+#     task :prepare => :environment do
+#       Rake::Task["db:seed"].invoke
+#     end
+#   end
+# end
