+++ app/controllers/application_controller.rb
@@ -2,10 +2,24 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-  include SessionsHelper
 
+  def log_in(user)
+    session[:user_id] = user.id
+  end
 
-  # <% if logged_in? -%>Welcome, <%= current_user.name %><% end -%>
+  def current_user
+    @current_user ||= User.find_by(id: session[:user_id])
+  end
 
+  def logged_in?
+    current_user != nil
+  end
+
+  def log_out
+    session.delete(:user_id)
+    @current_user = nil
+  end
+
+  helper_method :log_in, :current_user, :logged_in?, :log_out
 
 end
