+++ app/views/sessions/new.html.erb
@@ -1,4 +1,5 @@
 <h1>Log in!</h1>
+<%= @errors %>
 
 <div class="login">
     <%= form_for(:session, url: login_path) do |f| %>
