+++ app/views/users/_form.html.erb
@@ -12,12 +12,12 @@
 			</ul>
 		</div>
 	<% end %>
- 
+
   <p>
     <%= f.label :name %>
     <%= f.text_field :name %>
   </p>
-  <p> 
+  <p>
     <%= f.label :email %>
     <%= f.email_field :email %>
   </p>
