+++ app/views/users/index.html.erb
@@ -1,1 +1,1 @@
-<h1>hihiihi</h1>
+<h1>Index page for users</h1>
