+++ tmp/capybara/capybara-201502051121132092726920.html
@@ -1,0 +1,62 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Scrapples</title>
+  <link data-turbolinks-track="true" href="/assets/application.css" media="all" rel="stylesheet" />
+  <script data-turbolinks-track="true" src="/assets/application.js"></script>
+  
+</head>
+<body>
+	<header>
+  <div class="container">
+    <nav>
+      <ul>
+          <li><a href="/login">Log in</a></li>
+      </ul>
+    </nav>
+  </div>
+</header>
+	<div class="container">
+		<div class="sign_up">
+  <h1>Sign up!</h1>
+
+  <form accept-charset="UTF-8" action="/users" class="new_user" id="new_user" method="post"><div style="display:none"><input name="utf8" type="hidden" value="&#x2713;" /></div>
+		<div id="error_explanation">
+		<h2>
+			1 error prohibited
+			this user from being saved:
+		</h2>
+		<ul>
+	  		<li>Name can&#39;t be blank</li>
+			</ul>
+		</div>
+ 
+  <p>
+    <div class="field_with_errors"><label for="user_name">Name</label></div>
+    <div class="field_with_errors"><input id="user_name" name="user[name]" type="text" value="" /></div>
+  </p>
+  <p> 
+    <label for="user_email">Email</label>
+    <input id="user_email" name="user[email]" type="email" value="raj@raj.com" />
+  </p>
+  <p>
+    <label for="user_password">Password</label>
+    <input id="user_password" name="user[password]" type="password" />
+  </p>
+  <p>
+    <label for="user_password_confirmation">Password confirmation</label>
+    <input id="user_password_confirmation" name="user[password_confirmation]" type="password" />
+  </p>
+
+  <p>
+    <input name="commit" type="submit" value="Create User" />
+  </p>
+ </form>
+ 
+  <a href="/users">profile</a>
+
+</div>
+
+	</div>
+</body>
+</html>
