+++ tmp/capybara/capybara-201502051140191182871459.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Scrapples</title>
+  <link data-turbolinks-track="true" href="/assets/application.css" media="all" rel="stylesheet" />
+  <script data-turbolinks-track="true" src="/assets/application.js"></script>
+  
+</head>
+<body>
+	<header>
+  <div class="container">
+    <nav>
+      <ul>
+          <li><a href="/users/8">Profile</a></li>
+          <li><a data-method="delete" href="/logout" rel="nofollow">Log out</a> </li>
+			</ul>
+      </ul>
+    </nav>
+  </div>
+</header>
+	<div class="container">
+		<h1>Welcome Raj!</h1>
+	</div>
+</body>
+</html>
