+++ tmp/capybara/capybara-201502051150501796785662.html
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>Scrapples</title>
+  <link data-turbolinks-track="true" href="/assets/application.css" media="all" rel="stylesheet" />
+  <script data-turbolinks-track="true" src="/assets/application.js"></script>
+  
+</head>
+<body>
+	<header>
+  <div class="container">
+    <nav>
+      <ul>
+          <li><a href="/login">Log in</a></li>
+      </ul>
+    </nav>
+  </div>
+</header>
+	<div class="container">
+		<h1>Log in!</h1>
+oops
+
+<div class="login">
+    <form accept-charset="UTF-8" action="/login" method="post"><div style="display:none"><input name="utf8" type="hidden" value="&#x2713;" /></div>
+
+      <label for="session_email">Email</label>
+      <input id="session_email" name="session[email]" type="email" />
+
+      <label for="session_password">Password</label>
+      <input id="session_password" name="session[password]" type="password" />
+
+      <input name="commit" type="submit" value="Log in" />
+</form>
+    <p>New user? <a href="/signup">Create a new account</a></p>
+</div>
+
+	</div>
+</body>
+</html>
